pyproject.toml
[project]
name = "final-event"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-genai>=1.56.0",
    "playwright>=1.57.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
